requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
helper==2.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
Theano==1.0.5
threadpoolctl==3.2.0
twilio==8.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
